requirements.txt
# Core
pandas==2.1.0
numpy==1.25.0
scipy==1.11.0
scikit-learn==1.3.0

# Data Models
pydantic==2.4.0

# Data I/O
openpyxl==3.1.2
pyarrow==13.0.0

# Visualization
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0

# Reporting
jinja2==3.1.2
xlsxwriter==3.1.9

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-benchmark==4.0.0
hypothesis==6.88.0

# Utilities
structlog==23.1.0
click==8.1.7
tqdm==4.66.0
